requirements.txt
python-telegram-bot==13.15
maigret==0.4.2
